// src/health/dto/health.response.dto.ts
/**
 * @Author: Gibeom Choi
 * @Date:   2023-06-06 16:14:22
 * @Last Modified by:   Gibeom Choi
 * @Last Modified time: 2023-06-08 17:13:07
 */
export class HealthResponseDTO {
  health_info_id: number;
  health_info_text?: string;
  health_info_blood_type?: string;
  health_info_medication?: string;
  health_info_height: number;
  health_info_weight: number;
}